isl_tab_basic_set_constrained_lexmin: allow conditional constraints
A conditional constraint is only activated if the previous
(necessarily optional) constraint has been disabled.
This allows the caller to specify several options that
are examined in turn.
The difference with a concatenation into a larger disjunctive constraint
is that the large disjunctive constraint is satisfied as soon as
one of the options is satisfied, while the cascading constraints
will force the solver to try and satisfy the first option
before considering the second option.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>